+++ config/symbols.us.stcen.txt
@@ -39,6 +39,7 @@ InitializeEntity = 0x80194F74;
 CollectHeart = 0x801958F4;
 CollectGold = 0x80195974;
 CollectSubweapon = 0x80195A50;
+CollectHeartVessel = 0x80195B68;
 EntityPrizeDrop = 0x80195C84;
 EntityExplosion = 0x801964F8;
 BlinkItem = 0x801965F4;
