+++ src/st/np3/394F0.c
@@ -819,20 +819,7 @@ void CollectSubweapon(u16 subWeaponIdx) {
     DestroyEntity(g_CurrentEntity);
 }
 
-void CollectHeartVessel(void) {
-    if (g_PlayableCharacter != PLAYER_ALUCARD) {
-        g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
-        g_Status.hearts += HEART_VESSEL_RICHTER;
-
-        if (g_Status.heartsMax < g_Status.hearts) {
-            g_Status.hearts = g_Status.heartsMax;
-        }
-    } else {
-        g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
-        g_api.func_800FE044(HEART_VESSEL_INCREASE, 0x4000);
-    }
-    DestroyEntity(g_CurrentEntity);
-}
+#include "../collect_heart_vessel.h"
 
 void CollectLifeVessel(void) {
     g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
